Added lighting hint system for calculating direct illumination
Other changes:
* Removed unused Attenuation class
* Added Projectable trait
* Implemented Projectable methods in Sphere class

+++ Code/SPath/src/kaloffl/spath/scene/materials/RefractiveMaterial.scala
@@ -4,7 +4,7 @@ import kaloffl.spath.math.Color
 
 object RefractiveMaterial {
 
-  def apply(color: Color, refractiveIndex: Float, glossiness: Double = 0) =
+  def apply(color: Color, refractiveIndex: Float, glossiness: Float = 0) =
     new Material(color, new RefractFunction(refractiveIndex, glossiness))
 
 }
